Fix SSM agent installation and enablement***
This commit fixes the installation and enablement of the SSM agent in the `bastion.tf` file. The `sudo systemctl enable amazon-ssm-agent` command was moved after the `sudo systemctl start amazon-ssm-agent` command to ensure proper enablement. Additionally, an extra line break was removed.

In the `vpces.tf` file, an extra line break was added after the `aws_vpc_endpoint` resource block.